Add files via upload
I couldn't upload the " .godot " folder with all it's content. But I don't think it is necessarily needed, because it is just the most minimal content possible. Everything is original Godot Script Template. The bare minimum for the 3D "character" exist with the most basic movement function.

In the next "patch" I will add the changes I did to the default Template so far, for the character to turn around with the camera. PS: This is my first time trying to create a proper repository, so I hope it's not too bad and I get feedback on how to improve it.